+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* **Breaking** Refactored how view-process config events notify. !!: TODO
+* **Breaking** Refactored how view-process config events notify.
     - Initial non default config state now reported as events on init.
     - All config and monitors info removed from `ViewProcessInitedArgs` and related API.
 * Fix  `VARS.animations_enabled` not updating when it is not set and the `sys_animations_enabled` changes.